requirements.txt
openai
groq
google-generativeai
mistralai

langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-mistralai
langchain-text-splitters

huggingface_hub
transformers
sentence-transformers

faiss-cpu
chromadb

tiktoken

python-dotenv
pydantic
pydantic-settings

fastapi
uvicorn

requests
httpx

numpy
pandas
tqdm

pypdf
python-docx
unstructured

torch
accelerate

streamlit
langchain
langchain-core
langchain-mistralai
pydantic
python-dotenv